% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{module_list}
\alias{module_list}
\title{List loaded modules}
\usage{
module_list()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
List currently loaded modules from your system's software stack
}
\author{
Chris Field <fieldc@ethz.ch>
}
