% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{module_load}
\alias{module_load}
\title{Load module}
\usage{
module_load(module)
}
\arguments{
\item{module}{The name of the module you want to load, as normally given to the 'module load' command}
}
\value{
None
}
\description{
Load a module from your system's software stack
}
\details{
modulesR does not strictly load any modules into the session that R is running in, but instead creates a hidden list of modules you have loaded or unloaded with the package's commands. The system() function is overwritten so that any modules are loaded before the requested command is executed. This is not efficient for a large number of system calls, where modules should be loaded ahead of R if possible.
}
\author{
Chris Field <fieldc@ethz.ch>
}
