% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{module_purge}
\alias{module_purge}
\title{Unload all modules}
\usage{
module_purge()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
Unload all modules from your system's software stack
}
\author{
Chris Field <fieldc@ethz.ch>
}
