% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{module_unload}
\alias{module_unload}
\title{Unload module}
\usage{
module_unload(module)
}
\arguments{
\item{module}{The name of the module you want to unload, as normally given to the 'module unload' command}
}
\value{
None
}
\description{
Unload a module from your system's software stack
}
\author{
Chris Field <fieldc@ethz.ch>
}
